requirements.txt
asgiref==3.8.1
Django==5.0.5
django-etc==1.4.0
django-hitcount==1.3.5
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1